Use vault auth method and connection helm stanzas for integration test

+++ test/integration/vaultpkisecret_integration_test.go
@@ -24,6 +24,9 @@ func TestVaultPKISecret(t *testing.T) {
 	testK8sNamespace := "k8s-tenant-" + testID
 	testPKIMountPath := "pki-" + testID
 	testVaultNamespace := ""
+	testVaultConnectionName := "vaultconnection-test-tenant-1"
+	testVaultAuthMethodName := "vaultauth-test-tenant-1"
+	testVaultAuthMethodRole := "role1"
 
 	clusterName := os.Getenv("KIND_CLUSTER_NAME")
 	require.NotEmpty(t, clusterName, "KIND_CLUSTER_NAME is not set")
@@ -40,6 +43,9 @@ func TestVaultPKISecret(t *testing.T) {
 			"k8s_test_namespace":       testK8sNamespace,
 			"k8s_config_context":       "kind-" + clusterName,
 			"vault_pki_mount_path":     testPKIMountPath,
+			"vault_connection_name":    testVaultConnectionName,
+			"vault_authmethod_name":    testVaultAuthMethodName,
+			"vault_authmethod_role":    testVaultAuthMethodRole,
 		},
 	}
 	if entTests := os.Getenv("ENT_TESTS"); entTests != "" {
@@ -57,45 +63,47 @@ func TestVaultPKISecret(t *testing.T) {
 	terraform.InitAndApply(t, terraformOptions)
 
 	crdClient := getCRDClient(t)
+	ctx := context.Background()
 
 	// Create a VaultConnection CR
-	testVaultConnection := &secretsv1alpha1.VaultConnection{
-		ObjectMeta: v1.ObjectMeta{
-			Name:      "vaultconnection-test-tenant-1",
-			Namespace: testK8sNamespace,
-		},
-		Spec: secretsv1alpha1.VaultConnectionSpec{
-			Address: testVaultAddress,
-		},
-	}
-
-	ctx := context.Background()
-	defer crdClient.Delete(ctx, testVaultConnection)
-	err := crdClient.Create(ctx, testVaultConnection)
-	require.NoError(t, err)
-
-	// Create a VaultAuth CR
-	testVaultAuth := &secretsv1alpha1.VaultAuth{
-		ObjectMeta: v1.ObjectMeta{
-			Name:      "vaultauth-test-tenant-1",
-			Namespace: testK8sNamespace,
-		},
-		Spec: secretsv1alpha1.VaultAuthSpec{
-			VaultConnectionRef: "vaultconnection-test-tenant-1",
-			Namespace:          testVaultNamespace,
-			Method:             "kubernetes",
-			Mount:              "kubernetes",
-			Kubernetes: &secretsv1alpha1.VaultAuthConfigKubernetes{
-				Role:           "role1",
-				ServiceAccount: "default",
-				TokenAudiences: []string{"vault"},
+	if !deployOperatorWithHelm {
+		testVaultConnection := &secretsv1alpha1.VaultConnection{
+			ObjectMeta: v1.ObjectMeta{
+				Name:      testVaultConnectionName,
+				Namespace: testK8sNamespace,
 			},
-		},
-	}
+			Spec: secretsv1alpha1.VaultConnectionSpec{
+				Address: testVaultAddress,
+			},
+		}
 
-	defer crdClient.Delete(ctx, testVaultAuth)
-	err = crdClient.Create(ctx, testVaultAuth)
-	require.NoError(t, err)
+		defer crdClient.Delete(ctx, testVaultConnection)
+		err := crdClient.Create(ctx, testVaultConnection)
+		require.NoError(t, err)
+
+		// Create a VaultAuth CR
+		testVaultAuth := &secretsv1alpha1.VaultAuth{
+			ObjectMeta: v1.ObjectMeta{
+				Name:      testVaultAuthMethodName,
+				Namespace: testK8sNamespace,
+			},
+			Spec: secretsv1alpha1.VaultAuthSpec{
+				VaultConnectionRef: testVaultConnectionName,
+				Namespace:          testVaultNamespace,
+				Method:             "kubernetes",
+				Mount:              "kubernetes",
+				Kubernetes: &secretsv1alpha1.VaultAuthConfigKubernetes{
+					Role:           testVaultAuthMethodRole,
+					ServiceAccount: "default",
+					TokenAudiences: []string{"vault"},
+				},
+			},
+		}
+
+		defer crdClient.Delete(ctx, testVaultAuth)
+		err = crdClient.Create(ctx, testVaultAuth)
+		require.NoError(t, err)
+	}
 
 	// Create a VaultPKI CR to trigger the sync
 	testVaultPKI := &secretsv1alpha1.VaultPKISecret{
@@ -104,7 +112,7 @@ func TestVaultPKISecret(t *testing.T) {
 			Namespace: testK8sNamespace,
 		},
 		Spec: secretsv1alpha1.VaultPKISecretSpec{
-			VaultAuthRef: "vaultauth-test-tenant-1",
+			VaultAuthRef: testVaultAuthMethodName,
 			Namespace:    testVaultNamespace,
 			Mount:        testPKIMountPath,
 			Name:         "secret",
@@ -119,7 +127,7 @@ func TestVaultPKISecret(t *testing.T) {
 	}
 
 	defer crdClient.Delete(ctx, testVaultPKI)
-	err = crdClient.Create(ctx, testVaultPKI)
+	err := crdClient.Create(ctx, testVaultPKI)
 	require.NoError(t, err)
 
 	// Wait for the operator to sync Vault PKI --> k8s Secret, and return the
